requirements.txt
pandas==2.2.3
pymongo==4.11
scikit-learn==1.6.1
numpy==1.26.4
scikit-surprise==1.1.4

